+++ src/vars/commandData.ts
@@ -7,8 +7,8 @@ const scoreListString =
     `Mods can be specified with +[mods], -mx [exact mods] or -me [exclude mods]
 The arguments \`pp\`, \`score\`, \`acc\`, \`bpm\` and \`miss\` use the following format:
 \`-key value\` to filter by that exact value (eg. -bpm 220)
-\`-key >value\` to filter scores below value (eg. -pp >500)
-\`-key <value\` to filter scores above value (eg. -acc <90)
+\`-key >value\` to filter scores above that value (eg. -pp >500)
+\`-key <value\` to filter scores below that value (eg. -acc <90)
 \`-key min..max\` to filter scores between min and max (eg. -miss 1..3)
 \`-key !value\` to filter scores that dont equal that value (eg. -bpm !200)
 The \`sort\` arg can be specified using -value (ie -recent)